+++ askbot/doc/source/changelog.rst
@@ -11,6 +11,7 @@ Development version
   the feature requires access to the server (Evgeny)
 * Added optional top banner to the question page (Evgeny)
 * Made "bootstrap mode" default and created instead "large site mode" (Evgeny)
+* Added interesting/ignored/subscribed tags to the user profile page (Paul Backhouse, Evgeny)
 
 0.7.43 (May 14, 2012)
 ---------------------